backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
torch==2.5.1
torchvision==0.20.1
Pillow==11.1.0
numpy==2.2.2
pydantic==2.10.5
pydantic-settings==2.7.1
aiofiles==24.1.0
